+++ CHANGELOG.md
@@ -3,11 +3,10 @@
 Canonical reference for changes, improvements, and bugfixes for cap.
 
 ## Next
+* fix (saml): always validate response and assertion signatures when signatures are present ([PR #180](https://github.com/hashicorp/cap/pull/180))
 
 ## 0.12.0
-
 * feat (oidc): add Claims for exposing provider server metadata ([PR #172](https://github.com/hashicorp/cap/pull/172))
-* fix (saml): always validate response and assertion signatures when signatures are present ([PR #180](https://github.com/hashicorp/cap/pull/180))
 
 ## 0.11.0